data_augmentation file splitted from utils
